+++ Api/Services/Invitations/IInvitationRecordService.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
-using HappyTravel.Edo.Api.Models.Agents;
 using HappyTravel.Edo.Api.Models.Invitations;
 using HappyTravel.Edo.Common.Enums;
 using HappyTravel.Edo.Data.Agents;
@@ -23,13 +21,5 @@ namespace HappyTravel.Edo.Api.Services.Invitations
         Task<Result<UserInvitation>> GetActiveInvitation(string code);
 
         UserInvitationData GetInvitationData(UserInvitation invitation);
-
-        Task<List<AgentInvitationResponse>> GetAgentAcceptedInvitations(int agentId);
-
-        Task<List<AgentInvitationResponse>> GetAgentNotAcceptedInvitations(int agentId);
-
-        Task<List<AgentInvitationResponse>> GetAgencyAcceptedInvitations(int agencyId);
-
-        Task<List<AgentInvitationResponse>> GetAgencyNotAcceptedInvitations(int agencyId);
     }
 }
